requirements.txt
# Multi-Agent Azure DevOps Automation System
# Python 3.9+

# Core dependencies
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Autogen framework (for Phase 2) - pinned to 0.2.x for stable API
pyautogen>=0.2.0,<0.3.0

# FastAPI and web server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Azure SDK packages
azure-identity>=1.15.0
azure-mgmt-compute>=30.0.0
azure-mgmt-storage>=21.0.0
azure-storage-blob>=12.19.0
azure-mgmt-resource>=23.0.0
azure-keyvault-secrets>=4.7.0

# Azure DevOps
azure-devops>=7.1.0b4

# Azure AI Search (for Phase 3 - RAG)
azure-search-documents>=11.4.0

# Azure OpenAI
openai>=1.10.0

# Azure Application Insights
applicationinsights>=0.11.10

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.26.0

# Logging
python-json-logger>=2.0.7

# Data validation
pydantic>=2.5.0

# Utilities
requests>=2.31.0
